Getting the task list to write to the database - not working
There may be something wrong with how the Task and TaskList objects are persisting in the helper classes. I'm trying to generate the id of the task list name in the TASK_LIST_ID column of the TASKS table to be able to tie the list of items to the table name when calling the createListServlet.